+++ Constraints/include/LieSpaceResidual.h
@@ -17,17 +17,17 @@ namespace LieSpaceResidual {
 
 // You have to make sure that fkPtr_ has already computed the corresponding forward kinematics
 // and its gradient or hessian before using this function!!!
-Eigen::Vector3d translationResidual(const std::unique_ptr<ForwardKinematicsSolver>& fkPtr_,
-                                    const Eigen::Vector3d& desiredPosition,
-                                    Eigen::MatrixXd* gradientPtr_ = nullptr,
-                                    Eigen::Array<Eigen::MatrixXd, 3, 1>* hessianPtr_ = nullptr);
+Eigen::Vector3f translationResidual(const std::unique_ptr<ForwardKinematicsSolver>& fkPtr_,
+                                    const Eigen::Vector3f& desiredPosition,
+                                    Eigen::MatrixXf* gradientPtr_ = nullptr,
+                                    Eigen::Array<Eigen::MatrixXf, 3, 1>* hessianPtr_ = nullptr);
 
 // You have to make sure that fkPtr_ has already computed the corresponding forward kinematics
 // and its gradient or hessian before using this function!!!
-Eigen::Vector3d rotationResidual(const std::unique_ptr<ForwardKinematicsSolver>& fkPtr_,
-                                 const Eigen::Matrix3d& desiredRotation,
-                                 Eigen::MatrixXd* gradientPtr_ = nullptr,
-                                 Eigen::Array<Eigen::MatrixXd, 3, 1>* hessianPtr_ = nullptr);
+Eigen::Vector3f rotationResidual(const std::unique_ptr<ForwardKinematicsSolver>& fkPtr_,
+                                 const Eigen::Matrix3f& desiredRotation,
+                                 Eigen::MatrixXf* gradientPtr_ = nullptr,
+                                 Eigen::Array<Eigen::MatrixXf, 3, 1>* hessianPtr_ = nullptr);
 
 }; // namespace LieSpaceResidual
 }; // namespace RAPTOR
