+++ Constraints/src/CircleSurfaceContactConstraints.cpp
@@ -35,12 +35,12 @@ void CircleSurfaceContactConstraints::compute(const VecX& z,
         const Vec3& rotation_torque = lambda.head(3);
         const Vec3& translation_force = lambda.tail(3);
 
-        const double contact_force      = translation_force(2) + csp.maxSuctionForce;
-        const double friction_force     = sqrt(pow(translation_force(0), 2) + pow(translation_force(1), 2));
-        const double ZMP_moment         = sqrt(pow(rotation_torque(0), 2) + pow(rotation_torque(1), 2));
+        const float contact_force      = translation_force(2) + csp.maxSuctionForce;
+        const float friction_force     = sqrt(pow(translation_force(0), 2) + pow(translation_force(1), 2));
+        const float ZMP_moment         = sqrt(pow(rotation_torque(0), 2) + pow(rotation_torque(1), 2));
 
-        const double max_friction_force = csp.mu * contact_force;
-        const double max_ZMP_moment     = csp.R * contact_force;
+        const float max_friction_force = csp.mu * contact_force;
+        const float max_ZMP_moment     = csp.R * contact_force;
 
         // (1) positive contact force
         g(i * 3 + 0) = contact_force;
